Restore the development directory
After my previous patch was merged, I noticed a problem I was not aware of.  This repo uses '..' in the .bowerrc file so that it installs dependencies in the PARENT of the repo directory.  This replaces the development directory, and also cd's to it so that the bower installed dependency files are not placed in a directory we did not create.