Add an example that shows an unintentionally bug
As I was reading the documentation, I wanted to experiment the possibilities that offers the observer feature.
And I noticed that I can change the property's value that is being observed. This leads to an infinity loop.
This is for me an obvious anti-pattern. Maybe will be useful for other developers to see this warning?