Maybe demonstrate one-way binding when appropriate
The example shows a div with two-way binding. Is this an opportunity to show when one-way binding would suffice?

P.S. I may be making an assumption that it's better to use one-way binding; and use two-way binding only when necessary. If that's not accurate: if two-way binding should be used unless one-way binding is needed, that would be helpful to know.