Clarify definition of custom property mixins
Mixins do not take an actual JS object, and the word "take" makes it seem more like a JS function that 'takes' a parameter so that definition is a little misleading.  Of course, the given example proves this to NOT be the case, but I think this is clearer anyway.